fix(connector): [AUTHORIZEDOTNET] widen psync reason-code fields that overflow u8
`SyncTransactionResponse::response_code` and `response_reason_code` were typed
`Option<u8>`, which caps at 255. Authorize.Net reason codes run well past that —
315 (invalid card number), 316 (invalid expiry), 318 (duplicate transaction) —
so any psync response carrying one failed to deserialize with
`invalid value: integer ..., expected u8`, producing exactly the same
`RESPONSE_DESERIALIZATION_FAILED` as an unmodelled `transactionStatus`.

Hyperswitch does not model either field, so it never fails on them. Neither field
is read anywhere in this connector; they are carried only so the raw values reach
`typed_connector_response`. Widening to `i32` preserves that while removing the
overflow.

This is the second failure mode behind the same shadow diffs addressed by the
`#[serde(other)]` catch-all, and had to be fixed for those to close:
juspay/hyperswitch-cloud#23192, #23193, #23195, #23196.

+++ crates/integrations/connector-integration/src/connectors/authorizedotnet/transformers.rs
@@ -2899,8 +2899,12 @@ pub struct SyncTransactionResponse {
     pub transaction_id: String,
     #[serde(rename = "transactionStatus")]
     pub transaction_status: SyncStatus,
-    pub response_code: Option<u8>,
-    pub response_reason_code: Option<u8>,
+    // Authorize.Net reason codes run past 255 (e.g. 315 invalid card number, 318
+    // duplicate transaction), so `u8` overflows and fails the whole psync body.
+    // Hyperswitch does not model these fields at all and so never fails on them;
+    // they are carried here only for `typed_connector_response`, never read.
+    pub response_code: Option<i32>,
+    pub response_reason_code: Option<i32>,
     pub response_reason_description: Option<String>,
     pub network_trans_id: Option<String>,
     // Additional fields available but not needed for our implementation
